kernel: use a defconfig format
As the CI scripts are using make olddefconfig for the builds,
it actually means that it is handling kconfig as if it is a
defconfig.

It means that:

- invalid/obsolete Kconfig symbols are suppressed;
- kernel defaults will be used for any new symbols.

Yet, a full .config file is harder to maintain, as we can't
really see what symbols will be used by make olddefconfig.
Also, it may keep listing symbols that were already removed
or replaced.

Instead, let's cleanup kconfig file, keeping just the contents
of "make savedefconfig".

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>