Add nodebug.fragment
This is a kernel config fragment intended to be merged with the
provided kernel config. Example for out-of-tree build:

	$ ./scripts/kconfig/merge_config.sh -m \
		/path/to/xe-ci/kernel/{kconfig,nodebug.fragment}
	$ mkdir build64
	$ mv .config build64/
	$ make O=build64 olddefconfig

Alternatively, this also works:

	$ mkdir build64
	$ KBUILD_OUTPUT=build64 ./scripts/kconfig/merge_config.sh \
		/path/to/xe-ci/kernel/{kconfig,nodebug.fragment}

The goal of this fragment is to disable options normally used by kernel
developers to debug the kernel that may create issues for UMD teams
trying to run benchmarks. By maintaining this as a fragment it's
possible to avoid a combinatorial explosion as more configs are desired.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>