kconfig: Use minimized distro config
Our kernel config comes from years of running CI without much structure
on how the config was updated. This replaces the current config by a
distro-based config so our results are more inline with what distros
will get.

This commit used a Ubuntu config as baseline, config-6.2.0-36-generic,
but any distro config could be used if in future we switch to another
one.

The following configs are added to match what we currently have:

	- 00-ci.fragment: contains the configurations not related to xe
	  module
	- 01-xe.fragment: contains the configuration related to xe

Note that these 2 fragments are already integrated in the kconfig file
after in this repo so the kconfig can be used standalone without having
to merge additional options.

Note that these recently introduced configs are not needed since they
are already part of the default:

	CONFIG_NUMA=y
	CONFIG_PGTABLE_LEVELS=5
	CONFIG_X86_5LEVEL=y
	CONFIG_DYNAMIC_MEMORY_LAYOUT=y

In order to keep the kernel slim the modules not needed by our
machines were removed by resorting to `make localmodconfig` on
ADL-P, MTL, LNL and machines used in our CI with the discrete cards PVC,
ATS-M and DG2. This should cover all the modules needed by our uses.
Assuming that /tmp/lsmod contain the files with lsmod outputs for each
machine, this is summarizes the procedure taken:

	# Copy the distro config (config-6.2.0-36-generic) to .config
	git clean -fxd
	cp /boot/config-6.2.0-36-generic .config

	# Bring config to our version
	make olddefconfig

	# Build the kernel normally and boot each target machine with
	# that kernel, then collect the output of lsmod and save all
	# of them to a directory. Using /tmp/lsmod/ here.

	# Use ADL-P as baseline:
	make LSMOD=/tmp/lsmod/adlp.txt localmodconfig
	cp .config .config.baseline

	# For each machine, add back the missing modules:
	>/tmp/missing.txt
	for f in /tmp/lsmod/{atsm,lnl,mtlu,pvc}.txt; do
		cat .config.baseline > .config;
		make LSMOD=$f localmodconfig 2>&1 >/dev/null | \
			tr ' ' '\n' | grep -e '^CONFIG' \
			>> /tmp/missing.txt;
	done;
	cat .config.baseline > .config;
	cat /tmp/missing.txt | sort -u | sed 's/^/-m /'  | xargs ./scripts/config

	# Merge back the desired fragments for CI and xe-related options
	./scripts/kconfig/merge_config.sh -m .config $XE_CI_PATH/kernel/00-ci.fragment
	./scripts/kconfig/merge_config.sh -m .config $XE_CI_PATH/kernel/01-xe.fragment

	# save defconfig to commit back to xe/ci
	make savedefconfig;
	cat defconfig > $XE_CI_PATH/kernel/kconfig

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>