Revert "kconfig: Use minimized distro config"
This reverts commit 9490c6cdd376ecec59626385d14e88e284b7845d
and commit 84df8d93d836595bc3238df233bd89c228893130.

After the mindistro config changes, CI was on fire with some machines
failing to boot. The second commit allows us to monitor the execution
and see how things improve, but it's better to revert the whole thing
and try again later.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>