hooks/W1: Stop disabling display
With https://patchwork.freedesktop.org/series/126598/, it's now possible
to build xe with display enabled when testing W=1 build.

If we need something like that again in future, we could probably
improve it by keeping the configs to be tweaked in an array, so we
handle them generically. Hopefully it's never needed again so KISS.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>