config: Use minimized distro config (try 2)
Try again to use a minimized distro config as in commit
84df8d93d836 ("kconfig: Use minimized distro config").

Main differences from that attempt:

	1) Run lsmod on more machines to gather used modules
	2) Enable some ethernet modules regardless of they not appearing
	   on the systems used
	3) CONFIG_FRAME_WARN is kept at 2048
	4) Enable again some debug configurations left out
	5) Squash 9490c6cdd376 ("kconfig: Bring some configs back to y")

Note that the approach taken this time to join the configs from several
machines was much simpler. The script
scripts/kconfig/streamline_config.pl, that is called by `make
localmodconfig` works fine if we just concatenate a bunch of text files
with the lsmod output on different machines.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>