config: Use minimized distro config (try 3)
Try again to use a minimized distro config as in commit
84df8d93d836 ("kconfig: Use minimized distro config") and commit
7c211481d07d ("config: Use minimized distro config (try 2)")

Now the config generation is scripted to make it easier to update.
It's also based on a more recent kernel (6.8-rc2).

This also splits more the fragments with the goal of eventually make it
useful for the i915 builds as well.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>