hooks: tone down 00-showenv
For CI executions, 00-showenv used to rely on a partial kernel checkout
so the git-log invocation wouldn't be too verbose. CI changed that to
properly support checkpatch so now it has a complete checkout.

While at it, it doesn't seem we need much of the stuff there: it's only
interesting to know what are the env vars CI is exporting so we can
tweak the next hooks.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>