Give me my console back
Stop removing simpledrm - it's more important to have a console working
than a single platform in CI that could be using a initcall_blacklist
or some smarter way in the tests to stop running tests in the wrong
card.

On a rc1 release where ethernet driver works 50% of the boot, I want to
at least confirm the machine is booting.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>