kernel/kconfig: Disable kernel panic on soft lockup
With that enabled, igt_runner is not able to collect and report soft 
lockups, with their associated call traces, when a kernel panic is
immediately triggered.  That gives useless 'incomplete' results with
no errors logged.  When disabled, igt_runner is able to capture those
error messages, detects kernel taint, and triggers an abort, which gives
much more informative reports.

Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>